+++ src/helpers/any.ts
@@ -1,4 +1,5 @@
-import type { ObjectS } from '../types';
+import type { ObjectT } from "../types";
 
-const any = <T extends ObjectS = ObjectS>(value?: T) => value as T;
+const any = <T extends ObjectT = ObjectT>(value?: T) =>
+  value as ObjectT extends T ? "any" : T;
 export default any;