+++ src/helpers/array.ts
@@ -1,8 +1,8 @@
-import { ARRAY } from '../constants';
-import type { ArrayCustom, NotReadonly, ObjectS } from '../types';
+import { ARRAY } from "../constants";
+import type { ArrayCustom, NotReadonly, ObjectT } from "../types";
 
-const array = <const T extends ObjectS>(value: T) => {
-  type Out = ArrayCustom<Extract<NotReadonly<T>, ObjectS>>;
+const array = <const T extends ObjectT>(value: T) => {
+  type Out = ArrayCustom<Extract<NotReadonly<T>, ObjectT>>;
 
   return { [ARRAY]: value } as Out;
 };