+++ src/helpers/optional.ts
@@ -1,7 +1,7 @@
-import { OPTIONAL } from '../constants';
-import type { __ObjectS, AnyArray, ArrayCustom, Optional } from '../types';
+import { OPTIONAL } from "../constants";
+import type { __ObjectT, AnyArray, ArrayCustom, Optional } from "../types";
 
-const optional = <T extends __ObjectS | ArrayCustom | AnyArray<__ObjectS>>(
+const optional = <T extends __ObjectT | ArrayCustom | AnyArray<__ObjectT>>(
   value?: T,
 ) =>
   ({