+++ src/helpers/partial.ts
@@ -1,7 +1,7 @@
-import { PARTIAL } from '../constants';
-import type { ObjectS, PartialCustom } from '../types';
+import { PARTIAL } from "../constants";
+import type { ObjectT, PartialCustom } from "../types";
 
-const partial = <T extends ObjectS>(value: T): T & PartialCustom => {
+const partial = <T extends ObjectT>(value: T): T & PartialCustom => {
   const entries = Object.entries(value).filter(([key]) => key !== PARTIAL);
   const out: any = {};