+++ src/helpers/soa.ts
@@ -1,8 +1,8 @@
-import { SOA } from '../constants';
-import type { NotReadonly, ObjectS, SoaCustom } from '../types';
+import { SOA } from "../constants";
+import type { NotReadonly, ObjectT, SoaCustom } from "../types";
 
-const soa = <const T extends ObjectS>(value: T) => {
-  type Out = SoaCustom<Extract<NotReadonly<T>, ObjectS>>;
+const soa = <const T extends ObjectT>(value: T) => {
+  type Out = SoaCustom<Extract<NotReadonly<T>, ObjectT>>;
   return { [SOA]: value } as Out;
 };